Add registrable decorator
